% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_seaducks.R
\name{add_seaducks}
\alias{add_seaducks}
\title{Append sea duck information to the Harvest Survey reference table}
\usage{
add_seaducks(ref_data)
}
\arguments{
\item{ref_data}{A reference data table}
}
\description{
The \code{add_seaducks} function includes the sea duck open date, close date, and bag limit for California and Washington in the reference table. The California and Washington sea duck seasons match the same dates and bags as the states' duck seasons.
}
\references{
\url{https://github.com/USFWS/migbirdMBHS}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
