% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bagdays.R
\name{bagdays}
\alias{bagdays}
\title{Summarize number of days hunted}
\usage{
bagdays(data, output = "table")
}
\arguments{
\item{data}{Daily data table}

\item{output}{Default is "table"
\itemize{
\item "table" - returns a table
\item "plot" - returns a ggplot
}}
}
\description{
The \code{bagdays} function determines the total number of days hunted per hunter and species group in the daily Harvest Survey data.
}
\references{
\url{https://github.com/USFWS/migbirdMBHS}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
