% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bagspp.R
\name{bagspp}
\alias{bagspp}
\title{Summarize total number of retrieved birds per hunter}
\usage{
bagspp(data, output = "table")
}
\arguments{
\item{data}{Daily data table}

\item{output}{Default is "table"
\itemize{
\item "table" - returns a table that reports the total number of species groups, the species group combination, and retrieved total per hunter
\item "n" - returns a treemap of the number of species groups per hunter
\item "species" - returns a treemap of the species group combinations represented in the data; each hunter’s pursued species are grouped for this visualization
\item "success" - excludes "retrieved = 0" records to return a treemap of the species group combinations that were successfully hunted; each hunter’s pursued species are grouped for this visualization
}}
}
\description{
The \code{bagspp} function summarizes daily Harvest Survey data to determine the total number of retrieved birds per hunter. The hunter’s data are further broken down to determine which species groups the hunter responded to harvesting.
}
\references{
\url{https://github.com/USFWS/migbirdHS}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
