% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctHS.R
\name{correctHS}
\alias{correctHS}
\title{Correct errors in the Harvest Survey data}
\usage{
correctHS(data)
}
\arguments{
\item{data}{Daily or season data table processed by \code{proofHS}}
}
\description{
The \code{correctHS} function processes the errors flagged in \code{proofHS} and returns a final data table.
}
\references{
\url{https://github.com/USFWS/migbirdMBHS}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
