% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huntdays.R
\name{huntdays}
\alias{huntdays}
\title{Summarize days hunted in Harvest Survey daily data}
\usage{
huntdays(data, type = "both", output = "table", average = TRUE)
}
\arguments{
\item{data}{Daily data table}

\item{type}{Default is "both"
\itemize{
\item "both" - returns days hunted summarized by state and species
\item "state" - returns days hunted summarized by state
\item "species" - returns days hunted summarized by species
}}

\item{output}{Default is "table"
\itemize{
\item "table" - returns a table
\item "plot" - returns a ggplot
}}

\item{average}{Default is TRUE
\itemize{
\item TRUE - returns average number of days
\item FALSE - returns total number of days
}}
}
\description{
The \code{huntdays} function summarizes the total number of days hunted in the daily data.
}
\references{
\url{https://github.com/USFWS/migbirdMBHS}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
