% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodata.R
\name{nodata}
\alias{nodata}
\title{Check for missing data}
\usage{
nodata(data, ref_data = NA, species = NA, report = FALSE)
}
\arguments{
\item{data}{Daily or season data table}

\item{ref_data}{The reference data table that corresponds to the year of the data}

\item{species}{Whether species should be included in the output, and if so, which species in particular:
\itemize{
\item NA - does not include species
\item "all" - includes all species for all states
\item A specific species group, to show all of the species group specified for all states, which may be one of:
\itemize{
\item "Band-tailed Pigeon", "Coots", "Ducks", "Gallinules", "Geese", "Rails", "Sandhill Crane", "Snipe", "Specially Regulated Sea Ducks", "Woodcock"}
}}

\item{report}{Is this function being used in the R markdown season report? Defaults to FALSE.}
}
\description{
The \code{nodata} function checks to see if any states or species are missing from the daily or season Harvest Survey online data.
}
\references{
\url{https://github.com/USFWS/migbirdMBHS}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
