% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openclose.R
\name{openclose}
\alias{openclose}
\title{Check for out-of-season harvest}
\usage{
openclose(data, ref_data, state = NA, summary = F)
}
\arguments{
\item{data}{Daily data table}

\item{ref_data}{The reference data table that corresponds to the year of the daily data}

\item{state}{Default is NA, but can be used to specify what state should be reported in the output. State values can be one of:
\itemize{
\item "Alabama", "Alaska", "Arizona", "Arkansas", "California", "Colorado", "Connecticut", "Delaware", "Florida", "Georgia", "Idaho", "Illinois", "Indiana", "Iowa", "Kansas", "Kentucky", "Louisiana", "Maine", "Maryland", "Massachusetts", "Michigan", "Minnesota", "Mississippi", "Missouri", "Montana", "Nebraska", "Nevada", "New Hampshire", "New Jersey", "New Mexico", "New York", "North Carolina", "North Dakota", "Ohio", "Oklahoma", "Oregon", "Pennsylvania", "Rhode Island", "South Carolina", "South Dakota", "Tennessee", "Texas", "Utah", "Vermont", "Virginia", "Washington", "West Virginia", "Wisconsin", "Wyoming"
}}

\item{summary}{Defaults to FALSE
\itemize{
\item TRUE - returns a table summarized by state and species group to report the type and count of out-of-season values
\item FALSE - returns a table unsummarized data
}}
}
\description{
The \code{openclose} function checks if harvest dates fall outside of a state's and species's season in the daily data.
}
\references{
\url{https://github.com/USFWS/migbirdHS}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
