% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overbags.R
\name{overbags}
\alias{overbags}
\title{Check for overbags in the daily data}
\usage{
overbags(data, ref_data, state = NA, over = T, summary = F)
}
\arguments{
\item{data}{Daily data table}

\item{ref_data}{The reference data table that corresponds to the year of the daily data}

\item{state}{Default is NA, but can be used to specify what state should be reported in the output. State values can be one of:
\itemize{
\item "Alabama", "Alaska", "Arizona", "Arkansas", "California", "Colorado", "Connecticut", "Delaware", "Florida", "Georgia", "Idaho", "Illinois", "Indiana", "Iowa", "Kansas", "Kentucky", "Louisiana", "Maine", "Maryland", "Massachusetts", "Michigan", "Minnesota", "Mississippi", "Missouri", "Montana", "Nebraska", "Nevada", "New Hampshire", "New Jersey", "New Mexico", "New York", "North Carolina", "North Dakota", "Ohio", "Oklahoma", "Oregon", "Pennsylvania", "Rhode Island", "South Carolina", "South Dakota", "Tennessee", "Texas", "Utah", "Vermont", "Virginia", "Washington", "West Virginia", "Wisconsin", "Wyoming"
}}

\item{over}{Defaults to TRUE
\itemize{
\item TRUE - returns only records with retrieved values that exceed the bag limit
\item FALSE - returns a table of all values
}}

\item{summary}{Defaults to FALSE
\itemize{
\item TRUE - returns a table summarized by state and species to report the max, mean, min, and count of overbag values
\item FALSE - returns a table unsummarized data
}}
}
\description{
The \code{overbags} function checks for retrieved values over the bag limit per state and species in the Harvest Survey daily data.
}
\references{
\url{https://github.com/USFWS/migbirdHS}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
