% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overdays.R
\name{overdays}
\alias{overdays}
\title{Check for overdays in the daily or season data}
\usage{
overdays(data, ref_data)
}
\arguments{
\item{data}{Daily or season data table}

\item{ref_data}{The reference data table that corresponds to the year of the data}
}
\description{
The \code{overdays} function checks Harvest Survey daily or season data to determine if total days hunted exceed the length of the state and species' season.
}
\references{
\url{https://github.com/USFWS/migbirdMBHS}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
