% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partyhunt.R
\name{partyhunt}
\alias{partyhunt}
\title{Check for group hunts in the daily data}
\usage{
partyhunt(data)
}
\arguments{
\item{data}{Daily data table}
}
\description{
The \code{partyhunt} function checks to see if any comments in the daily Harvest Survey data indicate that the bag value reported was for two or more people.
}
\references{
\url{https://github.com/USFWS/migbirdMBHS}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
