% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partyproof.R
\name{partyproof}
\alias{partyproof}
\title{Adjust group hunts reported in Harvest Survey daily data}
\usage{
partyproof(data, ref_data, outpath)
}
\arguments{
\item{data}{Daily data table}

\item{ref_data}{Reference data table for the same year as the daily data}

\item{outpath}{Path to write the change log .csv; must end with a forward slash}
}
\description{
The \code{partyproof} function checks for, and allows the user to edit, any group hunts in the daily Harvest Survey data. Comments are parsed to determine party size. Adjustments to number of birds retrieved must be entered manually, and if deemed necessary, the retrieved value will be divided by the party size. A report file is written out as a .csv to record all changes made.
}
\references{
\url{https://github.com/USFWS/migbirdMBHS}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
