% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proofHS.R
\name{proofHS}
\alias{proofHS}
\title{Check for errors in daily and season data}
\usage{
proofHS(data, ref_data)
}
\arguments{
\item{data}{Daily or season data table}

\item{ref_data}{The reference data table that corresponds to the year of the daily data}
}
\description{
The \code{proofHS} function checks for overbag and overdays values in the Harvest Survey season data. In addition to overbag and overdays, daily data are checked for early and late hunts.
}
\references{
\url{https://github.com/USFWS/migbirdHS}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
