% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refcheck.R
\name{refcheck}
\alias{refcheck}
\title{Check for errors in the reference table}
\usage{
refcheck(ref_data)
}
\arguments{
\item{ref_data}{A reference data table}
}
\description{
The \code{refcheck} function checks for issues in the all_seasons Harvest Survey reference table.
}
\references{
\url{https://github.com/USFWS/migbirdMBHS}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
