% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responselag.R
\name{responselag}
\alias{responselag}
\title{Visualize response lag}
\usage{
responselag(data, type = "count")
}
\arguments{
\item{data}{Daily data table}

\item{type}{Default is "count"
\itemize{
\item "count" - returns the distribution of response lag
\item "lag" - returns the relationship between time lag and number of birds retrieved
\item "date" - returns the relationship between response date and harvest date
}}
}
\description{
The \code{responselag} function creates plots to visualize the amount of time between hunt date and response date in daily Harvest Survey data.
}
\references{
\url{https://github.com/USFWS/migbirdHS}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
