% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieved.R
\name{retrieved}
\alias{retrieved}
\title{Summarize number of birds retrieved in Harvest Survey data}
\usage{
retrieved(data, type = "both", output = "table", average = TRUE)
}
\arguments{
\item{data}{Daily or season data table}

\item{type}{Default is "both"
\itemize{
\item "both" - returns birds retrieved summarized by state and species
\item "state" - returns birds retrieved summarized by state
\item "species" - returns birds retrieved summarized by species
}}

\item{output}{Default is "table"
\itemize{
\item "table" - returns a table
\item "plot" - returns a ggplot
}}

\item{average}{Default is TRUE
\itemize{
\item TRUE - returns average birds retrieved
\item FALSE - returns total birds retrieved
}}
}
\description{
The \code{retrieved} function summarizes the total number of birds retrieved in the daily or season data.
}
\references{
\url{https://github.com/USFWS/migbirdHS}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
