% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrievedmap.R
\name{retrievedmap}
\alias{retrievedmap}
\title{Map mean retrieved birds per species and state}
\usage{
retrievedmap(data)
}
\arguments{
\item{data}{Daily data table}
}
\description{
The \code{retrievedmap} function creates a hex bin map of the United States from daily Harvest Survey data. Maps are plotted in a grid to display all species. States that are red have the highest mean harvest, states that are yellow have the lowest mean harvest, and blank (or white) states have no data.
}
\references{
\url{https://github.com/USFWS/migbirdMBHS}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
