% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submitted.R
\name{submitted}
\alias{submitted}
\title{Check if records were submitted}
\usage{
submitted(data, type = "totals", output = "table")
}
\arguments{
\item{data}{Daily or season data table}

\item{type}{"totals" for a summary of all Y and N responses, "state" to view Y and N responses by state, or "species" to view Y and N responses by species group}

\item{output}{"table" or "plot"}
}
\description{
The \code{submitted} function checks the has_submitted field in either Harvest Survey daily or season totals data.
}
\references{
\url{https://github.com/USFWS/migbirdHS}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
