% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_ref.R
\name{wrangle_ref}
\alias{wrangle_ref}
\title{Consistent reference data wrangling}
\usage{
wrangle_ref(ref_data)
}
\arguments{
\item{ref_data}{The reference data table that corresponds to the year of the data}
}
\description{
The internal \code{wrangle_ref} function edits the reference data table consistently among all functions in the migbirdHS package.
}
\references{
\url{https://github.com/USFWS/migbirdHS}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
