% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_seaducks.R
\name{add_seaducks}
\alias{add_seaducks}
\title{Add sea duck dates for eligible states}
\usage{
add_seaducks(ref_data)
}
\arguments{
\item{ref_data}{The reference data table that corresponds to the year of the daily data}
}
\description{
The \code{add_seaducks} function amends the ref_data for AK, CT, DE, ME, MD, MA, NH, NJ, NY, OR, RI, VA, WA, and CA by adding lines to account for seaduck seasons; it uses the same values as duck harvest for open, close, bag, and possession.
}
\references{
\url{https://github.com/USFWS/migbirdMBHS}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
