% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportHS.R
\name{reportHS}
\alias{reportHS}
\title{Write an R markdown report from template}
\usage{
reportHS(inpath, partypath, type, year, outpath, file)
}
\arguments{
\item{inpath}{File path to the folder containing the daily, season, and reference Harvest Survey .csv files}

\item{partypath}{File path to the folder containing results from partyproof function}

\item{type}{Type of report. One of the following options may be supplied:
\itemize{
\item survey_analytics - for a summary report of survey analytics
\item season_report - for a summary of the Harvest Survey data}}

\item{year}{The hunting season for which Harvest Survey data were collected}

\item{outpath}{Folder in which to save the completed report}

\item{file}{What the report file should be named}
}
\description{
Create documentation with figures and tables that summarizes Harvest Survey daily and season data.
}
\references{
\url{https://github.com/USFWS/migbirdHarvestData}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
